server/requirements.txt
aiohttp==3.10.3
databases==0.9.0
discord.py==2.6.4
httpx==0.28.1
openpyxl==3.1.5
pycryptodome==3.20.0
python-socketio==5.16.0
Requests==2.32.5
SQLAlchemy==2.0.37
starlette==0.50.0
uvicorn==0.40.0
